Logo, UI, Player, URP
Modified logo to have it more style friendly 32x32.
Modified the UI in a prefab for later use.
Reorganized player script.
Added Controls class for Player.
Added controls to player as a serializable field, now you can use it for showing it in the UI dynamically even change it (but don't)